add delete not allowed exception

+++ CWM_Django_part2/storefront2/store/views.py
@@ -55,6 +55,8 @@ def product_detail(request, id):
         serializer.save()
         return Response(serializer.data)
     elif request.method == 'DELETE':
+        if product.orderitems.count() > 0:
+            return Response(status=status.HTTP_405_METHOD_NOT_ALLOWED)
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
